# openApi.yml
openapi: 3.0.0
info:
  title: Event Booking API
  version: 1.0.0
  description: A Go-powered API for booking events
  contact:
    name: API Support
    url: http://www.example.com/support
    email: support@example.com
...
servers:
  - url: http://localhost:8080/v1/api
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /events/{id}/register:
    post:
      description: Register user for an event
      tags:
        - events
      operationId: registerUser
      
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: User registered for the event
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      eventId:
                        type: string
                      userId:
                        type: string

    delete:
      description: Cancel registration for an event
      tags:
        - events
      operationId: cancelRegistration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Registration cancelled
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...